separate model from mixin
